[IMP] diprem: Add new sequence for all the Sale Orders associated
A new sequence was created for the  Sale Orders created from the opportunities
so that after the first one is generated, it's number don't change and add a
suffix like the following ' - REV001', ' - REV002', ' - REV003'  and so on,
to all of the quotations associated to the same opportunity.
(e.g. 'SO001', 'SO001 - REV001', 'SO001 - REV002')

Changes were made in the Diprem Repo.